+++ docs/source/releasenotes.rst
@@ -11,6 +11,10 @@ Upcoming release
   - Add built-in wait period for all locators, instead of failing immediately
   - Add preview images for matched locators in Robot Framework log
 
+- Library **Cloud.Google**:
+
+  - Add support for Apps Script service
+  - Add support for Drive service
 
 6.7.2
 -----