+++ docs/source/releasenotes.rst
@@ -25,6 +25,8 @@ Upcoming release
   - Add keywords to add/remove labels from GMail messages
   - Add keyword ``Do Message Actions`` for performing custom set of actions on selected messages
 
+- Library **Tables**: Correctly handle empty fields when filtering
+
 6.7.2
 -----