+++ docs/source/releasenotes.rst
@@ -29,6 +29,8 @@ Upcoming release
   - RPA.Browser.Playwright has been added, enabling use of playwright based robotframework-browser library
   - RPA.Browser was moved to RPA.Browser.Selenium, and the old import RPA.Browser is kept as deprecated alias for now
 
+- Library **Tables**: Correctly handle empty fields when filtering
+
 6.7.2
 -----